[Keyboard] add dztech/dz65rgb/v3 keyboard (#13017)
