+++ keyboards/dztech/dz65rgb/v2/readme.md
@@ -1,0 +1,16 @@
+# DZ65RGB V2
+
+A customizable 65% RGB keyboard.
+
+* Keyboard Maintainer: [moyi4681](https://github.com/moyi4681)
+* Hardware Supported: DZ65RGB V2
+* Hardware Availability: [KBDfans](https://kbdfans.com/)
+
+## Making firmware
+
+Make example for this keyboard (after setting up your build environment):
+
+    make dztech/dz65rgb/v2:default
+
+See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
+
